--- src/components/App.module.css
@@ -1,11 +1,0 @@
-.App {
-    display: flex;
-    justify-content: center;
-}
-
-.page {
-    flex-direction: column;
-    max-width: 1920px;
-    margin-left: auto;
-    margin-right: auto;
-}
