+++ src/components/App/App.module.css
@@ -1,0 +1,33 @@
+.App {
+    display: flex;
+    justify-content: center;
+}
+
+.page {
+    flex-direction: column;
+    max-width: 1920px;
+    margin-left: auto;
+    margin-right: auto;
+}
+
+.scrollContainer {
+    height: 755px;
+    overflow-y: scroll;
+}
+
+.scrollContainerSmall {
+    height: 464px !important;
+}
+
+.scrollContainer::-webkit-scrollbar {
+    width: 8px;
+}
+
+.scrollContainer::-webkit-scrollbar-track {
+    background: #2F2F37;
+}
+
+.scrollContainer::-webkit-scrollbar-thumb {
+    background-color: #8585AD;
+    border-radius: 10px;
+}
