+++ src/components/BurgerConstructor/BurgerConstructor.module.css
@@ -2,7 +2,6 @@
     display: flex;
     width: 568px;
     height: 812px;
-    overflow-y: scroll;
     flex-direction: column;
 }
 
@@ -21,16 +20,24 @@
 
 .productSection {
     display: flex;
-    align-items: center;
+    flex-direction: column;
+    align-items: end;
     justify-content: space-between;
+}
+
+.productElement {
+    min-width: 536px;
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
     margin-bottom: 10px;
 }
 
-.productSection:first-child {
-    justify-content: flex-end;
+.productElementWide {
+    min-width: 568px !important;
+    justify-content: space-between !important;
 }
 
 .productSection:last-child {
-    justify-content: flex-end;
     margin-bottom: 0;
 }
