--- src/components/BurgerIngredients.module.css
@@ -1,28 +1,0 @@
-.BurgerIngredients {
-    width: 600px;
-    height: 872px;
-    overflow-y: scroll;
-}
-
-.IngredientsItems {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-}
-
-.ProductTitle {
-    word-wrap: break-word;
-    text-align: center;
-    height: 48px;
-}
-
-.ProductPrice {
-    display: flex;
-    align-items: center;
-}
-
-.Product {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    position: relative;
-}
