+++ src/components/BurgerIngredients/BurgerIngredients.module.css
@@ -1,0 +1,3 @@
+.BurgerIngredients {
+    width: 600px;
+}
