+++ src/components/IngredientsItems/IngredientsItems.module.css
@@ -1,0 +1,4 @@
+.IngredientsItems {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+}
