+++ src/components/Product/Product.module.css
@@ -1,0 +1,17 @@
+.Product {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    position: relative;
+}
+
+.ProductPrice {
+    display: flex;
+    align-items: center;
+}
+
+.ProductTitle {
+    word-wrap: break-word;
+    text-align: center;
+    height: 48px;
+}
